Prepended "b" to the wind direction names to make them binary strings because the strings returned from TWI are also binary strings python3.
Added a try/except block around opening the serial port.

Default to five second polls of the station and prompt during configuration
to change that if desired.

The station may return barometric *pressure* but I believe most
(all?) users will "calibrate" their WR25 console to display the barometer
corrected for comparisons against sea level barometric pressure.  In weewx
terms, I changed the loop tag from "pressure" to "barometer" to make
sensible barometric readings display in the various skins that look for
"barometer".  This may not be technically correct but at 8258' elevation,
reporting as "pressure" caused weewx to calculate my "barometer" to >
41 inHg!

The WR25 reports a garbled string for below zero outdoor temperatures.
Instead of "-20F" it may report "0-20F".  This causes an exception when
we try to convert that string to a float.  Added code to reformat the
negative temperature if necessary.  I made this change was made and
published it to wview many years ago.